プレイヤーのアビリティとゴーレムの岩攻撃にカメラシェイクを追加

+++ Script/Object/Gimmick/GimmickDropRock.h
@@ -51,8 +51,8 @@ private:
 	static constexpr float WAIT_TIME = 3.0f;
 
 	//カメラシェイク関連
-	static constexpr int CAMERA_SHAKE_TIME = 0.2f;
-	static constexpr int CAMERA_SHAKE_POWER = 2.0f;
+	static constexpr float CAMERA_SHAKE_TIME = 0.2f;
+	static constexpr float CAMERA_SHAKE_POWER = 2.0f;
 
 	// パラメータ情報
 	ParameterGimmick* parameterDropRock_;
